switched to different tasks, moved main.cpp logic to class

+++ spider-servo/lib/mt6701/mt6701.cpp
@@ -106,7 +106,7 @@ bool MT6701::isCRCValid(uint32_t raw_data) {
   return crc == original_crc;
 }
 
-bool MT6701::read(double *angle, mt6701_status_t *field_status, bool *button_pushed, bool *track_loss ){
+bool MT6701::read(float *angle, mt6701_status_t *field_status, bool *button_pushed, bool *track_loss ){
   uint32_t data;
   uint8_t status;
   double _angle;
